account messaging interface
