+++ src/message/mod.rs
@@ -1,4 +1,5 @@
-#[allow(clippy::module_inception)]
 mod message;
+mod signed_message;
 
 pub use self::message::*;
+pub use self::signed_message::*;
